Add foundation for swapping card game variants from app and sidebar
