+++ react-solitaire/src/app/App.css
@@ -4,29 +4,3 @@
     text-align: center;
 }
 
-.top-row {
-    display: flex;
-    flex-direction: row;
-    padding-bottom: 1%;
-    gap: 2vw;
-}
-
-.row {
-    display: flex;
-    flex-direction: row;
-}
-
-.card-stacks {
-    justify-content: space-between;
-    gap: 2vw;
-}
-
-.deck-spacer {
-    width: 12.1rem;
-}
-
-.card-columns {
-    display: flex;
-    flex-direction: row;
-    gap: 2vw;
-}
