+++ react-solitaire/src/components/side-panel/SidePanel.css
@@ -78,3 +78,26 @@
     /* a clear border lets the background gradient shine through */
     border: 2px solid transparent;
 }
+
+.panel-dropdown {
+    width: 10rem;
+    height: 2rem;
+    font-family: 'Arial', sans-serif;
+    font-size: 1rem;
+    color: white;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.2);
+    margin: 0;
+    background-color: rgb(0, 0, 0, 0.5);
+    backdrop-filter: blur(10px);
+    border: solid 1px black;
+    border-radius: 5px;
+    cursor: pointer;
+    pointer-events: all;
+
+    text-align: center;
+}
+
+.card-game-group {    
+    width: 10rem;
+    margin-bottom: 2rem;
+}
