+++ react-solitaire/src/shared/solitaire-props.ts
@@ -1,0 +1,5 @@
+import { Variant } from "./variants";
+
+export interface SolitaireProps {
+    onVariantChanged: (variant: Variant) => void
+}
