+++ react-solitaire/src/shared/variants.ts
@@ -1,0 +1,4 @@
+export enum Variant {
+    Klondike = 'Klondike',
+    Freecell = 'Freecell'
+}
